Updated object constructors
/eve/object
---
- Created `/eve/object` folder
- Created `EveObject` base class
- Converted all effect constructors to es6/ classes (*all extends from `EveObject`)
- Rearranged method orders so they're all the same
- Updated method descriptions
- Fixed linting
- Moved all scratch variables to `EveObject.global`

/eve/object/EveMissile
---
- Combined `EveMissile` and `EveMissileWarhead` into one file
- Fixed a bug in `EveMissileWarhead` where the wrong vector was being updated

/eve/object/EveSpaceObject
---
- Added `HasLocatorPrefix` for finding locator prefixes
- Added `FindLocatorBonePrefix` for finding locator `Tw2Bone`s
- Added `FindLocatorsByPrefix` for finding locators by their prefix

/eve/EveLocator
- Added `atlasIndex0` property
- Added `atlasIndex1` property
- Added `bone` property

/eve/object/EveShip
---
- Removed `transform` array from `EveTurretSetLocatorInfo` and replaced with `locators` array. (*This allows developers to identify which transform belongs to what locator along with what it's name is*)
- Refactored `RebuildTurretSets` so that it assigns `Tw2Bone` to eve locators when they have them
- Refactored `RebuildBoosterSets` to use `FindLocatorsByPrefix` method
- Refactored `Update` so that only turret sets with bones get their transforms updated per frame
