+++ src/eve/EveLocator.js
@@ -1,13 +1,24 @@
 import {mat4} from '../math';
 
 /**
- * Contains transform information for Boosters, Turrets and XLTurrets
+ * Contains transform information for T3 Attachments, Boosters, Turrets and XLTurrets
+ *
+ * @param {string} [name='']
+ * @param {mat4} [transform=mat4.create()]
  * @property {string} name
  * @property {mat4} transform
- * @constructor
+ * @property {?number} atlasIndex0
+ * @property {?number} atlasIndex1
+ * @property {?Tw2Bone} bone
  */
-export function EveLocator()
+export class EveLocator
 {
-    this.name = '';
-    this.transform = mat4.create();
+    constructor(name='', transform=mat4.create())
+    {
+        this.name = name;
+        this.transform = mat4.clone(transform);
+        this.atlasIndex0 = null;
+        this.atlasIndex1 = null;
+        this.bone = null;
+    }
 }
