+++ src/eve/index.js
@@ -10,11 +10,11 @@ export * from './EveSpotlightSet';
 export * from './EveSpriteSet';
 export * from './EveTurretSet';
 export * from './EveSpaceObjectDecal';
-export * from './EvePlanet';
-export * from './EveShip';
-export * from './EveSpaceObject';
-export * from './EveTransform';
+export * from './object/EvePlanet';
+export * from './object/EveShip';
+export * from './object/EveSpaceObject';
+export * from './object/EveTransform';
 export * from './EveEffectRoot';
-export * from './EveMissile';
+export * from './object/EveMissile';
 export * from './EveSpaceScene';
 export * from './EveSOF';
